requirements.txt
aiodns==3.0.0
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Authlib==1.1.0
autopep8==2.0.0
backports.cached-property==1.0.2
beautifulsoup4==4.11.1
Brotli==1.0.9
cached-property==1.5.2
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
comtypes==1.1.14
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
docutils==0.19
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
exchangelib==4.8.0
fire==0.4.0
Flask==2.2.2
fonttools==4.37.4
fpdf2==2.5.7
frozenlist==1.3.3
furl==2.1.3
graphviz==0.13.2
greenlet==1.1.3
grpcio==1.49.1
grpcio-tools==1.48.1
h11==0.14.0
htmldocx==0.0.6
hubspot-api-client==4.0.6
idna==3.4
importlib-metadata==4.13.0
isodate==0.6.1
itsdangerous==2.1.2
java-access-bridge-wrapper==0.9.5
Jinja2==3.1.2
jsonpath-ng==1.5.3
jsonschema==4.16.0
lxml==4.9.1
MarkupSafe==2.1.1
mss==6.1.0
multidict==6.0.4
netsuitesdk==1.24.0
notifiers==1.3.3
ntlm-auth==1.5.0
numpy==1.23.4
O365==2.0.21
oauthlib==3.2.2
opencv-python-headless==4.5.5.64
openpyxl==3.0.10
orderedmultidict==1.0.1
outcome==1.2.0
overrides==7.3.1
packaging==21.3
pandas==1.5.1
pdfminer.six==20201018
Pillow==9.2.0
platformdirs==2.5.2
playwright==1.27.1
ply==3.11
protobuf==3.20.1
psutil==5.9.3
pycares==4.3.0
pycodestyle==2.9.1
pycparser==2.21
pyee==8.1.0
Pygments==2.13.0
pymssql==2.2.7
PyMySQL==1.0.2
pynput-robocorp-fork==4.0.0
pyotp==2.7.0
pyparsing==3.0.9
PyPDF2==1.28.6
pyperclip==1.8.2
pyrsistent==0.18.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-http-client==3.3.7
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinauto==0.6.8
PyYAML==5.4.1
redmail==0.4.2
requests==2.28.1
requests-file==1.5.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.0
robocorp-dialog==0.5.3
robotframework==5.0.1
robotframework-assertion-engine==0.6.0
robotframework-browser==14.1.0
robotframework-databaselibrary==1.2.4
robotframework-pythonlibcore==3.0.0
robotframework-requests==0.9.4
robotframework-sapguilibrary==1.1
robotframework-seleniumlibrary==6.0.0
robotframework-seleniumtestability==2.0.0
rpaframework==18.0.0
rpaframework-core==10.0.1
rpaframework-dialogs==4.0.0
rpaframework-pdf==5.0.0
rpaframework-recognition==5.0.0
rpaframework-windows==6.1.0
selenium==4.5.0
sendgrid==6.9.7
simple-salesforce==1.12.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.43
starkbank-ecdsa==2.1.0
stringcase==1.2.0
svg.path==6.2
tenacity==8.1.0
termcolor==2.0.1
tesseract==0.1.3
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tweepy==3.10.0
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
uiautomation==2.0.17
urllib3==1.26.12
webdriver-manager==3.8.4
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
yarl==1.8.2
zeep==4.1.0
zipp==3.9.0
